+++ CHANGELOG.md
@@ -46,6 +46,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Bug Fixes
 
+* (server) [#26585](https://github.com/cosmos/cosmos-sdk/pull/26585) lower default gRPC MaxSendMsgSize from ~2GB to 10MB
+
 ### Deprecated
 
 ## [v0.54.3](https://github.com/cosmos/cosmos-sdk/releases/tag/v0.54.3) - 2026-05-05